execution/requirements.txt
# Python dependencies for Veo execution scripts
# Install with: pip install -r requirements.txt

# API clients
anthropic>=0.39.0
requests>=2.31.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Async/concurrency (for batch operations)
aiohttp>=3.9.0

# Caching
redis>=5.0.0  # Optional: for distributed caching

# Logging
structlog>=24.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
